refactor(qbittorrent): move torrent progress into client module
